icmptunnel: Make step used to disable kernel ICMP replies optional
There are four known problems related to ICMP replies sent by kernel

  1) PACKET_DATA sent by client replied by kernel running server with
     copied payload that considered valid by client and delivered to
     tunnel device to client kernel network stack (i.e. sent data
     by client mirrored to it by server)

  2) In non-emulation mode with strict firewall between client and
     server that tracks ICMP sequence numbers in requests and replies
     punchthru will have no effect: replies will "close" sequence
     number window on firewall.

  3) Slight performance overhead on client side caused by processing
     two copies of data.

  4) Third-party applications may receive these ICMP datagrams that
     also might have affect on their behaviour/performance.

It is not possible to fix 1) in client code without introducing change
to communication protocol in compatible with previous versions way.

To make it without protocol header size increase with new field added we
reverse packet magic string for client replies to "LNUT" while keeping
it original for server. Both server and client now checks for new magic
values for request and replies that makes them incompatible with
previous versions.

For 3) overhead is nearly neligible or 4) it is up to such application
to filter these datagrams (e.g. BPF filter attached to socket).

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>